+++ CHANGELOG.md
@@ -8,3 +8,4 @@ Put the changelog BELOW the dashes. ANYTHING ABOVE IS IGNORED
 - Fixed a few tags that were still unfinished
 - Cherry particles are now much more responsive to wind
 - Fixed the Carnation's dye recipe being in the wrong group
+- The Oasis's foliage and grass colors and no longer as bright